Use SegRs to create/renew EERs.
Because the SegRs are available on the initiator and all stitch points,
use them as transport to talk to the next AS and receive the response.
Stitch points will have to use the next SegR instead of the current one.
Integration tests not passing yet due to lack of support for stitching,
will be fixed in posterior commits.