Paths only in the direction of the traffic.
All paths are created and stored in the direction of the traffic.
Whenever the initiator of a down-segr (being the last AS) uses that path
as a transport, it's reversed. This affects the store and the debug service.
Just use the stored segments at initiators and stitch points when
performing an EER operation.
Clone of a nil colibri path is nil.